requirements.txt
# Existing
starlette>=0.30.0
pydantic>=2
rich>=13
pytest>=8
torch>=2
numpy>=1
setuptools>=68

# API
fastapi>=0.109.0
uvicorn>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Docker
docker>=7.0.0

# Data
pandas>=2.0.0
pyarrow>=14.0.0

# Logging
loguru>=0.7.0

# Settings
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

bittensor